Deprecate various methods of the KpointsData class
Some methods of the KpointsData to analyze the Bravais lattice
and set k-point paths based on that analysis, contain bugs for
more complicated cell structures like monoclinic and triclinic
cells. A proper implementation can be found in the Seekpath
package. In prepartion of moving towards using this package, we
deprecate the broken methods and move them to a legacy file in
the aiida.tools.data.array.kpoints folder. For the time being
the original methods will still refer to the original implementation
with a deprecation warning that will be emitted.
